% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_lm.R
\name{bayes_lm}
\alias{bayes_lm}
\title{Conjugate update in Bayesian linear model}
\usage{
bayes_lm(formula, data, m, C, phi)
}
\arguments{
\item{formula}{A formula}

\item{data}{A dataframe}

\item{m}{Mean in prior distribution of beta}

\item{C}{Variance in prior distribution of beta}

\item{phi}{Conditional variance of y given beta is phi I.}
}
\value{
An object of class 'blm'
}
\description{
Conjugate update in Bayesian linear model
}
\examples{
form <- dist ~ speed + I(speed^2)
mm <- lm(form, data=cars)
coef(summary(mm))
sigma(mm)

bb <- bayes_lm(form, data=cars, m=c(1, 1, 1), C=diag(10,3), phi=2)
bb

coef(bb)
sigma(bb)

if (require(ggplot2)){
pp <- qplot(speed, dist, data=cars)
pp + geom_line(aes(x=speed, y=predict(bb)), col="red") +
     geom_line(aes(x=speed, y=predict(mm)), col="blue")
}
}
\author{
Søren Højsgaard
}
