% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmBayes-package.R
\docType{package}
\name{lmBayes}
\alias{lmBayes}
\alias{lmBayes-package}
\title{lmBayes}
\description{
This package implements Bayesian linear regression.
}
\author{
Samuel I. Berchuck \email{sib2@duke.edu}
}

