% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_lm.R
\name{predint}
\alias{predint}
\title{Prediction interval}
\usage{
predint(object, level = 0.95, nsim = 100, newdata, ...)
}
\arguments{
\item{object}{A 'blm' object}

\item{level}{Level for interval}

\item{nsim}{Number of simulations to base the interval on}

\item{newdata}{New dataframe}

\item{...}{Additional arguments, currently unused.}
}
\description{
Prediction interval
}
\author{
Søren Højsgaard
}
