% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmbayes_main.R
\name{lmb}
\alias{lmb}
\title{Conjugate update in Bayesian linear model}
\usage{
lmb(formula., data., prior, update.sd = TRUE)
}
\arguments{
\item{formula.}{A formula}

\item{data.}{A dataframe}

\item{prior}{named list m (Mean in prior distribution of beta), C (Variance in prior distribution of beta), sd (Conditional variance of y given beta is sd I)}

\item{update.sd}{Should conditional variance be updated (emperical Bayes)}
}
\value{
An object of class 'blm'
}
\description{
Conjugate update in Bayesian linear model
}
\examples{
form <- dist ~ speed + I(speed^2)
mm <- lm(form, data=cars)
coef(summary(mm))
sigma(mm)

bb <- lmb(form, data=cars, prior=list(m=c(1, 1, 1), C=diag(10,3), sd=2))
bb

coef(bb)
sigma(bb)

if (require(ggplot2)){
pp <- qplot(speed, dist, data=cars)
pp + geom_line(aes(x=speed, y=predict(bb)), col="red") +
     geom_line(aes(x=speed, y=predict(mm)), col="blue")
}
}
\author{
Søren Højsgaard
}
